+++ lib/raven/client.rb
@@ -21,6 +21,12 @@ module Raven
     def send_event(event, hint = nil)
       return false unless configuration.sending_allowed?(event)
 
+      event = configuration.before_send.call(event, hint) if configuration.before_send
+      if event.nil?
+        configuration.logger.info "Discarded event because before_send returned nil"
+        return
+      end
+
       # Convert to hash
       event = event.to_hash
 
@@ -32,9 +38,6 @@ module Raven
       configuration.logger.info "Sending event #{event[:event_id]} to Sentry"
 
       content_type, encoded_data = encode(event)
-
-      event = configuration.before_send.call(event, hint) if configuration.before_send
-      return false if event.nil?
 
       begin
         transport.send_event(generate_auth_header, encoded_data,