+++ lib/raven/instance.rb
@@ -120,10 +120,10 @@ module Raven
             configuration.async.call(evt.to_json_compatible)
           rescue => ex
             logger.error("async event sending failed: #{ex.message}")
-            send_event(evt, obj)
+            send_event(evt, make_hint(obj))
           end
         else
-          send_event(evt, obj)
+          send_event(evt, make_hint(obj))
         end
         Thread.current["sentry_#{object_id}_last_event_id".to_sym] = evt.id
         evt
@@ -216,6 +216,10 @@ module Raven
           capture_type(exception, options)
         end
       end
+    end
+
+    def make_hint(obj)
+      obj.is_a?(String) ? { :exception => nil, :message => obj } : { :exception => obj, :message => nil }
     end
   end
 end