+++ spec/raven/integration_spec.rb
@@ -55,4 +55,51 @@ RSpec.describe "Integration tests" do
     @instance.capture_exception(build_exception)
     expect(@io.string).to match(/OK!$/)
   end
+
+  describe '#before_send' do
+    it "change event before sending (capture_exception)" do
+      @stubs.post('/prefix/sentry/api/42/store/') { [200, {}, 'ok'] }
+
+      @instance.configuration.server = 'http://12345:67890@sentry.localdomain/prefix/sentry/42'
+      @instance.configuration.before_send = lambda { |event, hint|
+        expect(hint[:exception]).not_to be nil
+        expect(hint[:message]).to be nil
+        event.environment = 'testxx'
+        event
+      }
+
+      event = @instance.capture_exception(build_exception)
+      expect(event.environment).to eq('testxx')
+
+      @stubs.verify_stubbed_calls
+    end
+
+    it "change event before sending (capture_message)" do
+      @stubs.post('/prefix/sentry/api/42/store/') { [200, {}, 'ok'] }
+
+      @instance.configuration.server = 'http://12345:67890@sentry.localdomain/prefix/sentry/42'
+      @instance.configuration.before_send = lambda { |event, hint|
+        expect(hint[:exception]).to be nil
+        expect(hint[:message]).not_to be nil
+        expect(event.message).to eq('xyz')
+        event.message = 'abc'
+        event
+      }
+
+      event = @instance.capture_message('xyz')
+      expect(event.message).to eq('abc')
+
+      @stubs.verify_stubbed_calls
+    end
+
+    it "return nil" do
+      @instance.configuration.server = 'http://12345:67890@sentry.localdomain/prefix/sentry/42'
+      @instance.configuration.before_send = lambda { |_event, _hint|
+        nil
+      }
+
+      @instance.capture_exception(build_exception)
+      expect(@instance.client.transport).to receive(:send_event).exactly(0)
+    end
+  end
 end